[Compose] Cleanup and support Visibility in DSL
Can now set Visible/Invisible/Gone on a widget when using the
ConstraintLayout DSL.
